Update how samples are calculated
Because the thread that plays the pattern(s) has millisecond resolution, the tempo is not exact. When rendering, instead of calculating the number of sample needed using a straight mathematical calculation based on time, modify to take into account the same calculation used when playing. This results in a .wav file that loops correctly.